Move main screen components into main screen folder
