+++ src/components/main-screen/bottom-bar/bottom-bar.component.scss
