+++ src/components/main-screen/bottom-bar/bottom-bar.component.ts
@@ -2,11 +2,11 @@ import { HttpErrorResponse } from '@angular/common/http';
 import { Component, OnDestroy } from '@angular/core';
 import { Subscription, timer } from 'rxjs';
 
-import { NotificationType, PrinterState, PrinterStatus, TemperatureReading } from '../../model';
-import { ConfigService } from '../../services/config.service';
-import { EnclosureService } from '../../services/enclosure/enclosure.service';
-import { NotificationService } from '../../services/notification.service';
-import { SocketService } from '../../services/socket/socket.service';
+import { NotificationType, PrinterState, PrinterStatus, TemperatureReading } from '../../../model';
+import { ConfigService } from '../../../services/config.service';
+import { EnclosureService } from '../../../services/enclosure/enclosure.service';
+import { NotificationService } from '../../../services/notification.service';
+import { SocketService } from '../../../services/socket/socket.service';
 
 @Component({
   selector: 'app-bottom-bar',
