+++ src/components/main-screen/job-status/height-progress/height-progress.component.scss
