+++ src/components/main-screen/job-status/job-status.component.html
