+++ src/components/main-screen/print-control/print-control.component.html
